+++ constable/vs.c
@@ -79,7 +79,7 @@ void vs_fill(vs_t *to)
  */
 void vs_invert(vs_t *to)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		to[i] = ~(to[i]);
 }
 
@@ -90,7 +90,7 @@ void vs_invert(vs_t *to)
  */
 void vs_set(const vs_t *from, vs_t *to)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		to[i] = from[i];
 }
 
@@ -101,7 +101,7 @@ void vs_set(const vs_t *from, vs_t *to)
  */
 void vs_add(const vs_t *from, vs_t *to)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		to[i] |= from[i];
 }
 
@@ -113,7 +113,7 @@ void vs_add(const vs_t *from, vs_t *to)
  */
 void vs_sub(const vs_t *from, vs_t *to)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		to[i] &= ~(from[i]);
 }
 
@@ -127,7 +127,7 @@ void vs_sub(const vs_t *from, vs_t *to)
  */
 void vs_mask(const vs_t *from, vs_t *to)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		to[i] &= from[i];
 }
 
@@ -139,7 +139,7 @@ void vs_mask(const vs_t *from, vs_t *to)
 int vs_test(const vs_t *test, const vs_t *vs)
 {
 	/* ATTENTION, see also object_cmp_vs() in object.c */
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		if (vs[i] & test[i])
 			return 1;
 	return 0;
@@ -147,7 +147,7 @@ int vs_test(const vs_t *test, const vs_t *vs)
 
 int vs_issub(const vs_t *subset, const vs_t *set)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		if ((set[i] & subset[i]) != subset[i])
 			return 0;
 	return 1;
@@ -160,7 +160,7 @@ int vs_issub(const vs_t *subset, const vs_t *set)
  */
 int vs_isclear(const vs_t *vs)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		if (vs[i])
 			return 0;
 	return 1;
@@ -173,7 +173,7 @@ int vs_isclear(const vs_t *vs)
  */
 int vs_isfull(const vs_t *vs)
 {
-	for (int i = 0; i < VS_WORDS; i++)
+	for (size_t i = 0; i < VS_WORDS; i++)
 		if (vs[i] != ~((vs_t)0))
 			return 0;
 	return 1;
